fix(#1395): BLAKE3 Rust acceleration — fix hash_fast.py fallback chain (#1453)
- Add blake3 = "1.5" dep to rust/nexus_fast/Cargo.toml
- Add hash_content() and hash_content_smart() to rust/nexus_fast/src/lib.rs
- Fix hash_fast.py: make blake3 import conditional with try/except,
  derive _PYTHON_BLAKE3_AVAILABLE from actual import result
- Add warning log when Rust acceleration unavailable
- Fix cross-platform hash bug: cast usize → u64 in to_le_bytes()
  for consistent hashes between Rust and Python on 32/64-bit
- Add 22 unit tests covering hash format, consistency, edge cases,
  sampling behavior, backend detection, and fallback chain

+++ src/rust/lib.rs
@@ -600,8 +600,10 @@ fn hash_content_smart(content: &[u8]) -> String {
         // Last 64KB
         hasher.update(&content[content.len() - SAMPLE_SIZE..]);
 
-        // Also include file size to differentiate files with same samples
-        hasher.update(&content.len().to_le_bytes());
+        // Also include file size to differentiate files with same samples.
+        // Cast to u64 for cross-platform consistency with Python's
+        // len(content).to_bytes(8, byteorder="little").
+        hasher.update(&(content.len() as u64).to_le_bytes());
 
         hasher.finalize().to_hex().to_string()
     }
